src/main/java/edu/usc/csci310/project/exception/GlobalExceptionHandler.java
package edu.usc.csci310.project.exception;

import org.springframework.stereotype.Component;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.http.ResponseEntity;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.MethodArgumentNotValidException;
import org.springframework.validation.FieldError;
import java.util.HashMap;
import java.util.Map;

@Component("GlobalExceptionHandler")
@ControllerAdvice
public class GlobalExceptionHandler {
    // // exception handlers
    // @ExceptionHandler(InvalidCredentialsException.class)
    // public ResponseEntity<?> handleInvalidCredentials(InvalidCredentialsException ex) {
    //     return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(ex.getMessage());
    // }

    @ExceptionHandler(UsernameExistsException.class)
    public ResponseEntity<?> handleUsernameExistsException(UsernameExistsException ex) {
        return ResponseEntity.badRequest().body(ex.getMessage());
    }

    @ExceptionHandler(MethodArgumentNotValidException.class)
    public ResponseEntity<?> handleValidationExceptions(MethodArgumentNotValidException ex) {
        // Extract and return the validation errors
        Map<String, String> errors = new HashMap<>();
        ex.getBindingResult().getAllErrors().forEach(error -> {
            String fieldName = ((FieldError) error).getField();
            String errorMessage = error.getDefaultMessage();
            errors.put(fieldName, errorMessage);
        });
        return ResponseEntity.badRequest().body(errors);
    }

}
